Use BigDecimal

+++ app/src/main/java/xyz/kewiany/menusy/common/OrderedProductsData.kt
@@ -1,9 +1,10 @@
 package xyz.kewiany.menusy.common
 
 import xyz.kewiany.menusy.domain.model.OrderedProduct
+import java.math.BigDecimal
 
 data class OrderedProductsData(
     val products: List<OrderedProduct>,
     val totalQuantity: Int,
-    val totalPrice: Float,
+    val totalPrice: BigDecimal,
 )
