+++ app/src/main/java/xyz/kewiany/menusy/domain/model/HistoryProduct.kt
@@ -1,12 +1,13 @@
 package xyz.kewiany.menusy.domain.model
 
 import xyz.kewiany.menusy.presentation.utils.HistoryProductUiItem
+import java.math.BigDecimal
 
 data class HistoryProduct(
     val productId: Long = 0,
     val name: String,
     val description: String,
-    val price: Float,
+    val price: BigDecimal,
     val quantity: Int
 )
